requirements.txt
Django>=3.26,<3.30
djangorestframework>=3.12.4,<3.13.0
python-dotenv>=0.19.0,<0.20.0
pyjwt>=2.1.0,<2.2.0